# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Scaling API
  description: API for the social media scaling application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /:
    get:
      summary: Health check
      description: Returns a simple greeting message
      operationId: getRoot
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: Hello chi!

  /auth/signup:
    post:
      summary: Sign up
      description: Register a new user account
      operationId: signup
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Login
      description: Authenticate user and return access token
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout
      description: Invalidate the current access token
      operationId: logout
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me:
    get:
      summary: Get current user
      description: Retrieve the authenticated user's information
      operationId: getMe
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me/feed:
    get:
      summary: Get news feed
      description: Retrieve tweets from users that the authenticated user follows (pull-based news feed with cursor-based pagination)
      operationId: getFeed
      tags:
        - feed
      security:
        - bearerAuth: []
      parameters:
        - name: count
          in: query
          description: Number of tweets to return (default 20, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Cursor for pagination (timestamp). Omit for first page, use next_cursor from previous response for next page.
          required: false
          schema:
            type: string
            format: date-time
            example: "2024-01-15T10:30:00.123456Z"
      responses:
        '200':
          description: Feed retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}:
    get:
      summary: Get user
      description: Retrieve a user by ID
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}/follow:
    put:
      summary: Follow user
      description: Follow the specified user
      operationId: followUser
      tags:
        - follows
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Followed successfully
        '400':
          description: Invalid request (e.g., cannot follow yourself)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Unfollow user
      description: Unfollow the specified user
      operationId: unfollowUser
      tags:
        - follows
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Unfollowed successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}/followers:
    get:
      summary: Get followers
      description: Get list of users who follow the specified user
      operationId: getFollowers
      tags:
        - follows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of followers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}/followees:
    get:
      summary: Get followees
      description: Get list of users the specified user is following
      operationId: getFollowees
      tags:
        - follows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of followees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tweets:
    get:
      summary: List tweets
      description: Retrieve a paginated list of tweets
      operationId: listTweets
      tags:
        - tweets
      parameters:
        - name: count
          in: query
          description: Number of tweets to return (default 20, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Cursor for relative pagination. Use next_cursor or previous_cursor from the response to navigate pages.
          required: false
          schema:
            type: integer
        - name: max_id
          in: query
          description: Returns tweets with an ID less than or equal to the specified ID (absolute position)
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tweets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TweetsResponse'

    post:
      summary: Create tweet
      description: Create a new tweet (max 255 characters)
      operationId: createTweet
      tags:
        - tweets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTweetRequest'
      responses:
        '201':
          description: Tweet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tweet'
        '400':
          description: Invalid request (e.g., content too long)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "0190a5e4-b890-7000-8000-000000000001"
        name:
          type: string
          example: johndoe
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
 
    SignupRequest:
      type: object
      properties:
        name:
          type: string
          example: johndoe
        password:
          type: string
          format: password
          example: secretpass123
      required:
        - name
        - password

    LoginRequest:
      type: object
      properties:
        name:
          type: string
          example: johndoe
        password:
          type: string
          format: password
          example: secretpass123
      required:
        - name
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      required:
        - user
        - token

    CreateTweetRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 255
          example: "Hello, world! This is my first tweet."
      required:
        - content

    Tweet:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "0190a5e4-b890-7000-8000-000000000010"
        user_id:
          type: string
          format: uuid
          example: "0190a5e4-b890-7000-8000-000000000001"
        content:
          type: string
          example: "Hello, world! This is my first tweet."
        likes_count:
          type: integer
          minimum: 0
          example: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - content
        - likes_count
        - created_at

    TweetWithUser:
      type: object
      description: Tweet with embedded user information (used in feed)
      properties:
        id:
          type: string
          format: uuid
          example: "0190a5e4-b890-7000-8000-000000000010"
        user_id:
          type: string
          format: uuid
          example: "0190a5e4-b890-7000-8000-000000000001"
        content:
          type: string
          example: "Hello, world! This is my first tweet."
        likes_count:
          type: integer
          minimum: 0
          example: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
      required:
        - id
        - user_id
        - content
        - likes_count
        - created_at
        - user

    TweetsResponse:
      type: object
      properties:
        tweets:
          type: array
          items:
            $ref: '#/components/schemas/Tweet'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - tweets
        - pagination

    FeedResponse:
      type: object
      description: News feed response with tweets from followed users
      properties:
        tweets:
          type: array
          items:
            $ref: '#/components/schemas/TweetWithUser'
        pagination:
          $ref: '#/components/schemas/FeedPagination'
      required:
        - tweets
        - pagination

    FeedPagination:
      type: object
      description: Cursor-based pagination for feed
      properties:
        count:
          type: integer
          description: Number of items returned in the current page
          example: 20
        next_cursor:
          type: string
          format: date-time
          nullable: true
          description: Cursor to use for the next page (timestamp). If null, this is the last page.
          example: "2024-01-15T10:30:00.123456Z"
      required:
        - count
        - next_cursor

    Pagination:
      type: object
      properties:
        count:
          type: integer
          description: Number of items returned in the current page
          example: 20
        cursor:
          type: integer
          description: Cursor specified in this request
          example: 20
        next_cursor:
          type: integer
          nullable: true
          description: Cursor to use for the next page. If null, this is the last page.
          example: 40
      required:
        - count
        - cursor
        - next_cursor

    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
      required:
        - users

    Error:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid request